added new features

+++ app/Repositories/V1/BookingRepository.php
@@ -2,14 +2,16 @@
 
 namespace App\Repositories\V1;
 
-use App\Models\V1\Booking;
-use App\Http\Resources\V1\BookingResource;
-use App\Models\User;
-use App\Models\V1\BookingPassengers;
 use App\Models\V1\Trip;
-use Illuminate\Container\Attributes\Log as AttributesLog;
+use App\Models\V1\Booking;
+use App\Models\UserBalance;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Log;
+use App\Models\V1\BookingPassengers;
+use App\Http\Resources\V1\BookingResource;
+use App\Models\BalanceTransaction;
+use App\Models\V1\CanceledBooking;
+use Illuminate\Support\Facades\Auth;
 
 class BookingRepository
 {
@@ -49,7 +51,6 @@ class BookingRepository
                     'message' => 'Trip not found'
                 ], 404);
             }
-
             $maxSeats = $trip->vehicle->seats;
             $requestedSeats = count($data['passengers']);
 
@@ -58,22 +59,70 @@ class BookingRepository
             }
 
             DB::beginTransaction();
+            if ($trip->available_seats < $requestedSeats) {
+                return response()->json(
+                    [
+                        'status' => 'error',
+                        'message' => 'Not enough seats available'
+                    ],
+                    422
+                );
+            }
 
+            $trip->available_seats = $trip->available_seats - $requestedSeats;
+            $trip->save();
+
+            if ($trip->available_seats <= 0) {
+                $trip->status = 'full';
+                $trip->save();
+            }
+
+            $totalPrice = $trip->price_per_seat * $requestedSeats;
+            $totalPrice = number_format((float)$totalPrice, 2, '.', '');
+            $userBalance = UserBalance::where('user_id', auth()->user()->id)->first();
+            if (!$userBalance) {
+                $userBalance = UserBalance::create([
+                    'user_id' => auth()->user()->id,
+                    'balance' => '100.00'
+                ]);
+            }
+
+            if ($userBalance->balance < $totalPrice) {
+                return response()->json(
+                    [
+                        'status' => 'error',
+                        'message' => 'Insufficient balance'
+                    ],
+                    422
+                );
+            }
             $booking = Booking::create([
                 'trip_id' => $data['trip_id'],
                 'user_id' => auth()->user()->id,
                 'seats_booked' => $requestedSeats,
-                'total_price' => $trip->price_per_seat * $requestedSeats,
+                'total_price' => $totalPrice,
+                'status' => 'confirmed',
                 'expired_at' => $trip->end_time
             ]);
 
-            $trip->available_seats -= $requestedSeats;
-            $trip->save();
+            if ($userBalance) {
+                $balancetranaction = new BalanceTransaction();
+                $balancetranaction->user_id = auth()->user()->id;
+                $balancetranaction->type = 'debit';
+                $balancetranaction->amount = $totalPrice;
+                $balancetranaction->balance_before = $userBalance->balance;
+                $balancetranaction->balance_after = $userBalance->balance - $totalPrice;
+                $balancetranaction->trip_id = $trip->id;
+                $balancetranaction->status = 'success';
+                $balancetranaction->reason =  'trip booking' . $booking->id . 'trip id' . $trip->id;
+                $balancetranaction->reference_id = $booking->id;
+                $balancetranaction->save();
 
-            if ($trip->available_seats <=    0) {
-                $trip->status = 'full';
-                $trip->save();
+                $userBalance->balance = ($userBalance->balance) - ($totalPrice);
+                $userBalance->save();
             }
+
+
 
             foreach ($data['passengers'] as $passenger) {
                 BookingPassengers::create([
@@ -89,14 +138,14 @@ class BookingRepository
         } catch (\Exception $e) {
             DB::rollBack();
             Log::error('Booking creation failed: ' . $e->getMessage());
-            return response()->json(['error' => 'Server error occurred please visit create booking section'], 500);
+            return response()->json(['status' => 'error', 'message' => 'Booking creation failed: ' . $e->getMessage()], 500);
         }
     }
 
 
     public function updateBooking($id, array $data)
     {
-        $booking = Booking::find($id);
+        $booking = Booking::where('user_id', auth()->user()->id)->find($id);
 
         if (is_null($booking)) {
             return response()->json($this->errorResponse, 404);
@@ -105,31 +154,42 @@ class BookingRepository
         try {
             DB::beginTransaction();
 
-            // Booking ma'lumotlarini yangilash
+            $requestedSeats = isset($data['passengers']) ? count($data['passengers']) : $booking->seats_booked;
+            //                                         2                       1
+            if ($requestedSeats > $booking->seats_booked) {
+                return response()->json(
+                    [
+                        'status' => 'error',
+                        'message' => 'you can not add more passengers than you have already booked'
+                    ],
+                    422
+                );
+            }
+
+            if ($requestedSeats == $booking->seats_booked) {
+                // user only want to update passengers
+                if (isset($data['passengers']) && is_array($data['passengers'])) {
+                    BookingPassengers::where('booking_id', $booking->id)->delete();
+                    foreach ($data['passengers'] as $passenger) {
+                        BookingPassengers::create([
+                            'booking_id' => $booking->id,
+                            'name' => $passenger['name'],
+                            'phone' => $passenger['phone'],
+                        ]);
+                    }
+                }
+            }
+
+
             $booking->update([
-                'trip_id' => $data['trip_id'] ?? $booking->trip_id,
-                'user_id' => $data['user_id'] ?? $booking->user_id,
-                'seats_booked' => isset($data['passengers']) ? count($data['passengers']) : $booking->seats_booked,
+                'seats_booked' => $requestedSeats,
                 'total_price' => isset($data['passengers'])
-                    ? $booking->trip->price * count($data['passengers'])
+                    ? $booking->trip->price_per_seat * count($data['passengers'])
                     : $booking->total_price,
                 'status' => $data['status'] ?? $booking->status,
             ]);
 
-            // Agar yangi yo‘lovchilar kelgan bo‘lsa — oldilarini o‘chirib, yangilarini qo‘shamiz
-            if (isset($data['passengers']) && is_array($data['passengers'])) {
-                // Oldingi yo‘lovchilarni o‘chirish
-                BookingPassengers::where('booking_id', $booking->id)->delete();
 
-                // Yangi yo‘lovchilarni yaratish
-                foreach ($data['passengers'] as $passenger) {
-                    BookingPassengers::create([
-                        'booking_id' => $booking->id,
-                        'name' => $passenger['name'],
-                        'phone' => $passenger['phone'],
-                    ]);
-                }
-            }
 
             DB::commit();
 
@@ -144,20 +204,105 @@ class BookingRepository
         }
     }
 
-    public function deleteBooking($id)
+    public function cancelBooking($bookingId)
     {
-        $booking = Booking::find($id);
-
-        if (is_null($booking)) {
-            return response()->json($this->errorResponse, 404);
-        }
+        DB::beginTransaction();
 
         try {
-            $booking->delete();
-            return response()->json($this->successResponse, 200);
-        } catch (\Exception $e) {
-            Log::error('Booking delete failed: ' . $e->getMessage());
-            return response()->json(['status' => 'error', 'message' => 'Server error occurred'],  500);
+            $user = Auth::user();
+
+            $booking = Booking::where('user_id', $user->id)->find($bookingId);
+            if (is_null($booking)) {
+                return response()->json([
+                    'status' => 'error',
+                    'message' => 'booking not found',
+                ]);
+            }
+
+            if ($booking->status == 'cancelled') {
+                return response()->json([
+                    'status' => 'error',
+                    'message' => 'booking already cancelled',
+                ], 422);
+            }
+
+            if ($booking->status == 'pending') {
+                return response()->json([
+                    'status' => 'error',
+                    'message' => 'Booking is not confirmed yet please wait',
+                ], 422);
+            }
+
+            if ($booking->status == 'completed') {
+                return response()->json([
+                    'status' => 'error',
+                    'message' => 'Booking is already completed, you can not cancel it',
+                ], 422);
+            }
+
+            $total = (float) $booking->total_price;
+
+            $userBalance = UserBalance::firstOrCreate(
+                ['user_id' => $user->id],
+                ['balance' => 0.00, 'currency' => 'som'] // kerak bo‘lsa currency
+            );
+
+            // 90% refund, 10% komissiya
+            $commission = round($total * 0.10, 2);
+            $refund = round($total - $commission, 2);
+
+            $balanceBefore = $userBalance->balance;
+            $userBalance->balance = $userBalance->balance + $refund;
+            $userBalance->save();
+            $balanceAfter = $userBalance->balance;
+
+            $trip = Trip::find($booking->trip_id);
+            if(is_null($trip)){
+                return response()->json([
+                    'status' => 'error',
+                    'message' => 'trip not found',
+                ]);
+            }
+            if($trip->total_seats < ($booking->seats_booked + $booking->trip->available_seats)){
+                return response()->json([
+                    'status' => 'error',
+                    'message' => 'Available seats should not be more than total seats',
+                ]);
+            }
+            $trip->available_seats = $trip->available_seats + $booking->seats_booked;
+            $trip->status = 'active';
+            $trip->save();
+
+            // BalanceTransaction
+            BalanceTransaction::create([
+                'user_id' => $user->id,
+                'type' => 'credit',
+                'amount' => $refund,
+                'balance_before' => $balanceBefore,
+                'balance_after' => $balanceAfter,
+                'trip_id' => $booking->trip_id,
+                'status' => 'success',
+                'reason' => 'Booking canceled - 90% refund ' . $refund . ' som, ' . ' ' . $commission . ' som commission, ' . $booking->id . ' booking id, ' . $trip->id . ' trip id, ',
+                'reference_id' => $booking->id,
+                'currency' => 'som',
+            ]);
+
+            // Booking status update
+            $booking->status = 'cancelled';
+            $booking->save();
+
+            DB::commit();
+
+            return response()->json([
+                'status' => 'success',
+                'message' => 'Booking canceled and refund issued.'
+            ]);
+        } catch (\Throwable $e) {
+            DB::rollBack();
+            return response()->json([
+                'status' => 'error',
+                'message' => 'Error: ' . $e->getMessage()
+            ], 500);
         }
     }
 }
